ci: remove unused "autoupdate" workflow (#2069)
This workflow auto-updates pull requests branches whenever changes
land on their destination branch. The workflow is triggered by assigning
a label to a pull request.

As far as I know, we are not using this workflow.

The workflow is running on Ubuntu 20, which is supported by GitHub until
April 1, 2025.

Since we are not using this workflow, I propose to not upgrade the
Ubuntu version, but instead to remove the entire workflow.

Signed-off-by: Miroslav Bajtoš <oss@bajtos.net>